# openapi.yaml
openapi: 3.1.1
info:
  title: CarConfigurator API
  version: 1.0.0
  description: >-
    Comprehensive car configuration API with internationalization, user management, and admin features. Fully compliant
    with OpenAPI 3.1.1 specification for Robot Framework training and automation testing.
  contact:
    name: CarConfigurator Team
    email: support@carconfigurator.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
components:
  schemas:
    Car:
      type: object
      required:
        - id
        - name
        - category
        - basePrice
        - description
      properties:
        id:
          type: string
          description: Unique car identifier
        name:
          type: string
          description: Car model name
          examples:
            - BMW X5
            - Audi Q7
            - Mercedes GLE
        category:
          type: string
          description: Car category
          examples:
            - SUV
            - Sedan
            - Coupe
        basePrice:
          type: number
          format: float
          description: Base price in Euro
          examples:
            - 45000
            - 52000
            - 68000
        description:
          type: string
          description: Car description
          examples:
            - Luxury SUV with latest technology
            - Premium sedan with hybrid drive
        imageUrl:
          type:
            - string
            - 'null'
          description: Car image URL (can be relative or absolute)
          examples:
            - /images/cars/elegance-sedan.jpg
            - https://example.com/car.jpg
        imageData:
          type:
            - string
            - 'null'
          description: Base64 encoded image data
        imageMimeType:
          type:
            - string
            - 'null'
          description: Image MIME type
          examples:
            - image/jpeg
            - image/png
            - image/webp
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
    Option:
      type: object
      required:
        - id
        - name
        - category
        - price
      properties:
        id:
          type: string
          description: Unique option identifier
        name:
          type: string
          description: Option name
          examples:
            - Lederausstattung
            - Sportmotor
            - Panorama-Schiebedach
        category:
          type: string
          description: Option category
          examples:
            - Innenausstattung
            - Motor
            - Lackierung
            - Felgen
        price:
          type: number
          format: float
          description: Option price in Euro
          examples:
            - 2500
            - 5000
            - 1200
        description:
          type:
            - string
            - 'null'
          description: Short option description
        detailedDescription:
          type:
            - string
            - 'null'
          description: Detailed option description
        imageUrl:
          type:
            - string
            - 'null'
          description: Option image URL (can be relative or absolute)
          examples:
            - /images/options/sport-package.jpg
            - https://example.com/option.jpg
        imageData:
          type:
            - string
            - 'null'
          description: Base64 encoded image data
        imageMimeType:
          type:
            - string
            - 'null'
          description: Image MIME type
        exclusiveGroup:
          type:
            - string
            - 'null'
          description: Exclusive group - options in the same group are mutually exclusive
        isRequired:
          type: boolean
          description: Indicates whether the option is required
          default: false
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
    User:
      type: object
      required:
        - id
        - email
        - name
        - role
      properties:
        id:
          type: string
          description: Unique user identifier
        email:
          type: string
          format: email
          minLength: 5
          description: User email address
        name:
          type: string
          minLength: 1
          description: User name
        role:
          type: string
          enum:
            - user
            - admin
          description: User role
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
    Configuration:
      type: object
      required:
        - id
        - name
        - totalPrice
        - userId
        - carId
      properties:
        id:
          type: string
          description: Unique configuration identifier
        name:
          type: string
          description: Configuration name
          examples:
            - Meine Traumkonfiguration
            - Sport-Edition
            - Comfort-Paket
        totalPrice:
          type: number
          format: float
          description: Total configuration price in Euro
        userId:
          type: string
          description: ID of the user who created the configuration
        carId:
          type: string
          description: ID of the configured car
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
        user:
          $ref: '#/components/schemas/User'
        car:
          $ref: '#/components/schemas/Car'
        options:
          type: array
          items:
            $ref: '#/components/schemas/ConfigurationOption'
    ConfigurationOption:
      type: object
      required:
        - configurationId
        - optionId
        - option
      properties:
        configurationId:
          type: string
          description: ID of the configuration
        optionId:
          type: string
          description: ID of the option
        option:
          $ref: '#/components/schemas/Option'
...
  securitySchemes:
    AdminAuth:
      type: apiKey
      in: cookie
      name: admin-session
      description: Admin-Session-Cookie für authentifizierte Anfragen
    UserAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT-Token für Benutzerauthentifizierung
paths:
...
  /api/configurations:
    get:
      tags:
        - Configurations
      summary: Get user configurations
      description: Retrieve all configurations for a specific user
      parameters:
        - in: query
          name: userId
          schema:
            type: string
          description: User ID to filter configurations
      responses:
        '200':
          description: Configurations retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    examples:
                      - true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Configuration'
        '500':
          description: Internal server error
    post:
      tags:
        - Configurations
      summary: Create a new configuration
      description: Create a new car configuration with selected options
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - carId
                - totalPrice
              properties:
                name:
                  type: string
                  examples:
                    - My Dream Car
                    - Sport Edition
                carId:
                  type: string
                  examples:
                    - car-1
                    - car-2
                totalPrice:
                  type: number
                  examples:
                    - 45000
                    - 67500
                userId:
                  type: string
                  examples:
                    - user-123
                selectedOptions:
                  type: array
                  items:
                    type: string
                  examples:
                    - - option-1
                      - option-2
      responses:
        '201':
          description: Configuration created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    examples:
                      - true
                  data:
                    $ref: '#/components/schemas/Configuration'
        '400':
          description: Invalid input data
        '500':
          description: Internal server error
...